Was added liquibase framework

+++ src/main/java/com/ssau/learn/dao/BankRepository.java
@@ -7,9 +7,8 @@ import java.rmi.server.UID;
 import java.util.Optional;
 
 public interface BankRepository extends JpaRepository<Bank, UID> {
-
     Optional<Bank> findBankByTerbankName(String terbankName);
     Optional<Bank> findBankByBik(String bik);
-    Optional<Bank> findById(int id);
-    void deleteById(int id);
+    Optional<Bank> findById(long id);
+    void deleteById(long id);
 }
